requirements.txt
# cloth segmentation
torch
torchvision
gradio
gdown
Pillow
opencv-python
numpy
pyyaml
tqdm
easydict==1.9.0
scikit-image
scikit-learn
opencv-python==4.8.0.76
tensorflow
joblib
matplotlib
pandas
albumentations==0.5.2
hydra-core==1.1.0
pytorch-lightning==1.2.9
tabulate
kornia==0.5.0
webdataset
packaging
wldhx.yadisk-direct
timm